Rename de controler registry a users
