+++ src/api/users.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const controller = require('../controller/registry');
+const controller = require('../controller/users');
 
 const router = express.Router();
